+++ graphql/schemas/Guild/customer.graphql
