+++ src/Baka/Traits/KanvasAppScopesTrait.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Baka\Traits;
+
+use Illuminate\Database\Eloquent\Builder;
+use Kanvas\Apps\Models\Apps;
+
+trait KanvasAppScopesTrait
+{
+    /**
+     * scopeApp.
+     *
+     * @param mixed $app
+     */
+    public function scopeFromApp(Builder $query, mixed $app = null): Builder
+    {
+        $app = $app instanceof Apps ? $app : app(Apps::class);
+
+        return $query->where('apps_id', $app->getId());
+    }
+}
